refactor: One error fixed, ensuring during CSV conversion that it checks if there are properties separated before fetching the value. And in Chunk.java, it iterates over the keys of the data, and if the value of the given key is numeric, it utilizes it for further calculations. BREAKING CHANGE: It's not fully functional, as it only displays the signs!

+++ sources/toolchain/rendering/src/main/java/codemetropolis/toolchain/rendering/model/BasicBlock.java
@@ -92,7 +92,7 @@ public class BasicBlock {
 		return stringId;
 	}
 
-	public int getShortId(){
+	public short getShortId(){
 		return shortId;
 	}
